[CHG] Adding AppStorage to sync states across the app + Project organization
