+++ DesignCodeiOS15V2/Components/BlobView.swift
