+++ DesignCodeiOS15V2/Navigation/ContentView.swift
@@ -1,0 +1,40 @@
+//
+//  ContentView.swift
+//  DesignCodeiOS15V2
+//
+//  Created by Damien Gautier on 08/03/2023.
+//
+
+import SwiftUI
+
+struct ContentView: View {
+    
+    @AppStorage("selectedTab") var selectedTab: Tab = .home
+    
+    var body: some View {
+        ZStack(alignment: .bottom) {
+            
+            Group {
+                switch selectedTab {
+                case .home:
+                    HomeView()
+                case .explore:
+                    AccountView()
+                case .notifications:
+                    AccountView()
+                case .library:
+                    AccountView()
+                }
+            }
+            .frame(maxWidth: .infinity, maxHeight: .infinity)
+            
+            TabBar()
+        }
+    }
+}
+
+struct ContentView_Previews: PreviewProvider {
+    static var previews: some View {
+        ContentView()
+    }
+}
