+++ DesignCodeiOS15V2/Navigation/TabBar.swift
@@ -7,48 +7,31 @@
 
 import SwiftUI
 
-struct TapBar: View {
+struct TabBar: View {
     
-    @State var selectedTab: Tab = .home
+    @AppStorage("selectedTab") var selectedTab: Tab = .home
     @State var color: Color = .teal
     @State var tabItemWidth: CGFloat = 0
     
     var body: some View {
-        ZStack(alignment: .bottom) {
-            
-            Group {
-                switch selectedTab {
-                case .home:
-                    ContentView()
-                case .explore:
-                    AccountView()
-                case .notifications:
-                    AccountView()
-                case .library:
-                    AccountView()
-                }
-            }
-            .frame(maxWidth: .infinity, maxHeight: .infinity)
-            
-            HStack {
-                buttons
-            }
-            .padding(.horizontal, 8)
-            .padding(.top, 14)
-            .frame(height: 88, alignment: .top)
-            .background(.ultraThinMaterial, in:
-                            RoundedRectangle(cornerRadius: 34, style: .continuous))
-            .background(
-                background
-            )
-            
-            .overlay(
-                overlay
-            )
-            .strokeStyle(cornerRadius: 34)
-            .frame(maxHeight: .infinity, alignment: .bottom)
-            .ignoresSafeArea()
+        HStack {
+            buttons
         }
+        .padding(.horizontal, 8)
+        .padding(.top, 14)
+        .frame(height: 88, alignment: .top)
+        .background(.ultraThinMaterial, in:
+                        RoundedRectangle(cornerRadius: 34, style: .continuous))
+        .background(
+            background
+        )
+        
+        .overlay(
+            overlay
+        )
+        .strokeStyle(cornerRadius: 34)
+        .frame(maxHeight: .infinity, alignment: .bottom)
+        .ignoresSafeArea()
     }
     
     var buttons: some View {
@@ -141,6 +124,6 @@ struct TapBar: View {
 
 struct TapBar_Previews: PreviewProvider {
     static var previews: some View {
-        TapBar()
+        TabBar()
     }
 }
