+++ DesignCodeiOS15V2/Views/AccountView.swift
@@ -64,7 +64,7 @@ struct AccountView: View {
     
     var menu: some View {
         Section {
-            NavigationLink(destination: ContentView()) {
+            NavigationLink(destination: HomeView()) {
                 Label("Settings", systemImage: "gear")
             }
             
@@ -72,7 +72,7 @@ struct AccountView: View {
                 Label("Billing", systemImage: "creditcard")
             }
             
-            NavigationLink { ContentView() } label: {
+            NavigationLink { HomeView() } label: {
                 Label("Help", systemImage: "questionmark")
             }
         }
